Refactor: batch 작업 분리
