+++ src/main/java/com/moyo/backend/batch/ranking/reader/GithubRankingRestClientImpl.java
@@ -1,4 +1,4 @@
-package com.moyo.backend.domain.temporary_batch.ranking.reader;
+package com.moyo.backend.batch.ranking.reader;
 
 import static com.moyo.backend.common.constant.MoyoConstants.GITHUB_QUERY_PAGING_SIZE;
 
@@ -12,9 +12,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 
-import com.moyo.backend.domain.temporary_batch.ranking.dto.GithubContributorDetails;
-import com.moyo.backend.domain.temporary_batch.ranking.dto.GithubRepoDetails;
-import com.moyo.backend.domain.temporary_batch.ranking.dto.RankingPreflight;
+import com.moyo.backend.batch.ranking.dto.GithubContributorDetails;
+import com.moyo.backend.batch.ranking.dto.GithubRepoDetails;
+import com.moyo.backend.batch.ranking.dto.RankingPreflight;
 
 @Slf4j
 @Component
